// api-app/src/common/enums/function-names.enum.ts
export enum FunctionNames {
  CreateNftBatch = 'createNFTBatch',
  MarketplaceTransferNft = 'marketplaceTransferNFT',
  SetPrice = 'setPrice',
  CreateProject = 'createProject',
  Vote = 'Vote',
  Proposal = 'createProposal',
  AllocationBreakdown = 'allocationBreakdown',
  TokenAllocation = 'tokenAllocation',
  SupportProject = 'supportProject'
}