// api-app/src/common/dto/common-parameter.dto.ts
export class CommonParameter {
  partnerId?: number;
  functionName?: string;
  adminAddress?: string;
  nftContractProxy?: string;
  kuFactoryAddress?: any;
  kumLogicAddress?: any;
  forwarder?: string;
  uuid?: string;
  transactionStatus?: string;
  tokenId?: any;
  message: any;
  transactionHash: any;
}